requirements.txt
https://github.com/New-dev0/Telethon/archive/master.zip
requests
cryptg
telethon-tgcrypto
